Structure cards into nested objects
I.e, there's an object for "you", an object for "primary", etc. This as
opposed to the previous `.you-email` previously used by the backend /
blockchain **AND** as opposed to the previous `[index].email` previously
assumed by the frontend. This decision was made to unify them and to
make the structure more human-friendly.