Move getClosestPerson to server; add frontend on /
Also tweaks to make it work well.

So now on / when you start entering contact info, the entire database
will be fuzzy searched, and the closest match, if it's good, will be
displayed, and you'll be asked if the person is relevant. If you select
it, the key of that person's entry (their original card) will be entered
in.