Completely nestle fabric within repository
This removes the dependency on fabric-samples

The chaincode is now named icecard

The problem ended up being that the package.json was still calling it
fabcar and trying to run a fabcar.js. I found a very convenient
environment variable that adds chaincode instantiation to the log, this
I clearly kept!

I'll move files around to be clean in a moment.